Issue #4587: Moved resources for FinalParameters check test
