+++ src/test/java/com/puppycrawl/tools/checkstyle/checks/FinalParametersCheckTest.java
@@ -33,7 +33,10 @@ import com.puppycrawl.tools.checkstyle.utils.CommonUtils;
 public class FinalParametersCheckTest extends BaseCheckTestSupport {
     @Override
     protected String getPath(String filename) throws IOException {
-        return super.getPath("checks" + File.separator + filename);
+        return super.getPath("checks" + File.separator
+                + "misc" + File.separator
+                + "finalparameters" + File.separator
+                + filename);
     }
 
     @Test
