+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/misc/finalparameters/InputFinalParametersReceiver.java
@@ -1,4 +1,4 @@
-package com.puppycrawl.tools.checkstyle.checks;
+package com.puppycrawl.tools.checkstyle.checks.misc.finalparameters;
 
 public class InputFinalParametersReceiver {
     public void foo4(InputFinalParametersReceiver this) {}
